refactor(rename): split rename/mod.rs god-file into submodules
